refactor: move common fixture to local pytest conftest file
